requirements.txt
Flask==2.3.3
pandas
azure-storage-blob
